Create README.md
A site needs to be designed where user will register either as a buyer or seller.
If registered as a seller then he will be able to advertise his property on the site mentioning the features of the site as location, area, price, additional features etc.
If registered as a buyer, he can go through the site and perform searches based on his requirement, take a look at various loan schemes and modes of payment available etc.
A user can also log-in to the system as a guest. A customer if using guest account will not have the permissions to advertise or search the properties but will be able to only enquire about the services that are provided by the company. He will be using a pre-designed Enquiry Form for this. This enquiry form will be mailed on the company’s mail-id.